campus-connect-ai/backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.6.1
SQLAlchemy==2.0.27
bcrypt==4.0.1
requests==2.32.3
pytest==7.2.2
Flask-Testing==0.8.1
firebase-admin==5.2.0
Flask-WTF==1.0.1
gunicorn==20.1.0
flask-dance==7.1.0
requests-oauthlib==2.0.0
oauthlib==3.2.2
PyMySQL==1.0.2
streamlit>=1.50.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.10.0
pandas>=2.1.0
beautifulsoup4>=4.12.2
diskcache>=5.5.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic
PyPDF2